requirements.txt
flask
numpy>=1.22
scipy
line-bot-sdk
gunicorn
deepface